Report Generation/requirements.txt
langchain-groq>=0.1.0
tavily-python>=0.3.1
pydantic>=2.5.0
python-dotenv>=1.0.0 